test of daos entitites

+++ src/main/java/dao/PersonDAO.java
@@ -28,7 +28,6 @@ public class PersonDAO {
             query.setParameter("hobbyName", hobbyName);
             return query.getResultList();
         }
-
     }
 
     public Person getPersonInfoByPhoneNumber(String number) {
@@ -49,13 +48,24 @@ public class PersonDAO {
         }
     }
 
-    public void deletePerson(Person person) {
+    /*public void deleteByPersonId(Integer id) {
         try (EntityManager em = emf.createEntityManager()) {
             em.getTransaction().begin();
-            em.remove(person);
+            em.remove(id);
             em.getTransaction().commit();
         }
+    }*/
+
+    public void deleteByPersonId(int personId) {
+        try(EntityManager em = emf.createEntityManager()) {
+            Person person = em.find(Person.class, personId);
+            if(person != null) {
+                em.remove(person);
+            }
+        }
+
     }
+
 
     public Person updatePerson(Person person) {
         try (EntityManager em = emf.createEntityManager()) {
@@ -66,11 +76,18 @@ public class PersonDAO {
         }
     }
 
+    public Person getPersonById(int Id) {
+        try (EntityManager em = emf.createEntityManager()) {
+            return em.find(Person.class, Id);
+        }
+    }
+
     public Person getPersonByName(String name) {
         try (EntityManager em = emf.createEntityManager()) {
-            return em.find(Person.class, name);
+            TypedQuery<Person> query = em.createQuery("SELECT p FROM Person p WHERE p.firstName = :name", Person.class);
+            query.setParameter("name", name);
+            return query.getSingleResult();
         }
-
     }
 
         public List<Person> findPersonsByCityZip(int zip) {
